Auth/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-jose[cryptography]
httpx

